pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "neuro-symbolic-healthcare"
version = "0.1.0"
description = "Production-ready neuro-symbolic MLOps system for healthcare AI with compliance checking"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Neuro-Symbolic Healthcare Team"}
]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Neural Components
    "transformers>=4.40.0",
    "sentence-transformers>=2.7.0",
    "torch>=2.2.0",
    "ollama>=0.3.0",
    
    # Symbolic Components
    "z3-solver>=4.12.0",
    "owlready2>=0.45",
    "rdflib>=7.0.0",
    
    # API & Web
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
    "httpx>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    
    # MLOps
    "mlflow>=2.12.0",
    "dvc>=3.50.0",
    
    # Monitoring
    "prometheus-client>=0.20.0",
    "structlog>=24.1.0",
    
    # Utilities
    "pyyaml>=6.0.1",
    "numpy>=1.26.0",
    "scipy>=1.13.0",
    "redis>=5.0.0",
    "aiofiles>=23.2.0",
    "tenacity>=8.2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "hypothesis>=6.100.0",
    "mypy>=1.9.0",
    "ruff>=0.3.0",
    "pre-commit>=3.6.0",
    "httpx>=0.27.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
]

[project.scripts]
nsh-server = "src.api.main:run_server"
nsh-test = "src.mlops.joint_testing:run_tests"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=term-missing"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "ANN", "B", "C4", "SIM"]
ignore = ["ANN101", "ANN102", "ANN401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]
